final refactoring of response objects
Got rid of the Accounts type classes as first class objects;
they now only serve as creators of arrays, so we don't have
to fake out an array. this way something like

    users = rest.get("/account/Naildrivin5/users")

returns an array of User.

However, I also monkeypatched Array to have success?
and not_modified? so that you don't have to check anything